+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Upgrade to imgui/cimgui 1.53.1
+  - Rename `Ui::show_test_window` to `Ui::show_demo_window`.
 
 
 ## [0.0.18] - 2017-12-23